requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
binwalk==2.3.3+5aebd7b
blinker==1.4
capstone==5.0.1
cassandra-driver==3.29.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloud-init==23.1.2
colorama==0.3.7
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cstruct==2.1
cycler==0.11.0
dataclasses==0.8
distlib==0.3.8
distro-info===0.18ubuntu0.18.04.1
filelock==3.4.1
Flask==2.0.3
geomet==0.2.1.post1
gnupg==2.3.1
html5lib==0.999999999
httplib2==0.9.2
humanfriendly==10.0
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
itsdangerous==2.0.1
jefferson==0.4.1
Jinja2==3.0.3
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
kafka-python==2.0.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
lxml==4.2.1
MarkupSafe==2.0.1
matplotlib==3.3.4
netifaces==0.10.4
numpy==1.19.5
oauthlib==2.0.6
olefile==0.45.1
PAM==0.4.2
pexpect==4.2.1
Pillow==8.4.0
platformdirs==2.4.0
psutil==5.9.6
psycopg2==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pycryptodome==3.19.0
PyGObject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==3.1.1
pyserial==3.4
python-apt==1.6.6
python-dateutil==2.8.2
python-debian==0.1.32
python-lzo==1.14
python-magic==0.4.16
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
sos==4.4
ssh-import-id==5.7
systemd-python==234
tk==0.1.0
Twisted==17.9.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
virtualenv==20.17.1
webencodings==0.5
Werkzeug==2.0.3
yaffshiv==0.1
zope.interface==4.3.2
